@@ -22,6 +22,7 @@
     "stylelint-library-report": "npx stylelint --custom-formatter 'node_modules/stylelint-junit-formatter' 'pattern-library/**/*.scss' --config pattern-library/src/scss/stylelint.config.js --fix > stylelint-library-results.xml",
     "eslint-theme": "npx eslint './js/*.js' --fix -f junit -o eslint-theme-results.xml",
     "eslint-library": "npx eslint './pattern-library/src/*.js' --fix -f junit -o eslint-library-results.xml",
+    "eslint-vue": "npx eslint './vue/**/*.vue'",
     "eslint-components": "npx eslint './pattern-library/components/**/*.js' --fix -f junit -o eslint-components-results.xml",
     "eslint": "npm run eslint-theme && npm run eslint-library && npm run eslint-components",
     "builder-up": "docker-compose --file storybook-compose.yml up -d builder",
@@ -85,7 +86,7 @@
     "browser-sync-webpack-plugin": "^2.3.0",
     "cross-env": "7.0.3",
     "dotenv-webpack": "^7.1.0",
-    "eslint": "^7.23.0",
+    "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-import": "^2.22.1",
@@ -94,6 +95,8 @@
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react": "^7.23.1",
     "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-plugin-vue": "^7.20.0",
+    "eslint-plugin-vuejs-accessibility": "^1.2.0",
     "faker": "^5.1.0",
     "glob": "^7.1.6",
     "jest": "^26.6.3",
